# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "B4 network packet processor REST API",
        "title": "B4 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/geoip": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeoIP"
                ],
                "summary": "List geoip categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GeoipResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GeoIP"
                ],
                "summary": "Add IP/CIDR blocks to a set",
                "parameters": [
                    {
                        "description": "CIDR blocks to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.AddGeoIpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AddIpResponse"
                        }
                    }
                }
            }
        },
...
        "/system/cache": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get cache statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/system/info": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Get system information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SystemInfo"
                        }
                    }
                }
            }
        },
        "/system/restart": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Restart the service",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.RestartResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.RestartResponse"
                        }
                    }
                }
            }
        },
        "/system/update": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Start update process",
                "parameters": [
                    {
                        "description": "Update request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.AddGeoIpRequest": {
            "type": "object",
            "properties": {
                "cidr": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "set_id": {
                    "type": "string"
                },
                "set_name": {
                    "type": "string"
                }
            }
        },
        "handler.AddIpResponse": {
            "type": "object",
            "properties": {
                "manual_cidrs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "total_cidrs": {
                    "type": "integer"
                }
            }
        },
...
        "handler.GeoipResponse": {
            "type": "object",
            "properties": {
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "handler.RestartResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "restart_command": {
                    "type": "string"
                },
                "service_manager": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "handler.SystemInfo": {
            "type": "object",
            "properties": {
                "arch": {
                    "type": "string"
                },
                "can_restart": {
                    "type": "boolean"
                },
                "is_docker": {
                    "type": "boolean"
                },
                "os": {
                    "type": "string"
                },
                "service_manager": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateRequest": {
            "type": "object",
            "properties": {
                "version": {
                    "type": "string"
                }
            }
        },
        "handler.UpdateResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "service_manager": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                },
                "update_command": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" to authorize",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}